Fix Val(...) when evaluating a multi module
We want to get the latest result that was produced. (Note that it is
likely that a more robust solution can be implemented when we re-do
outputs to be an object rather than hash of hash of array. I think
the sanity checks keep us from geting into any really odd situations
with this way, though.)