Simplify bad output error handling logic
We should only ever be calling `.get-output` on the top-level outputs
object if we don't think we have multi-instance data. Thus, doing so is
always a straight error. Throwing it from the fallback logic in an
instance outputs object will be caught and mapped appropriately to an
unset exception. The single-instance output object does *not* need to
try and figure out that situation; its attempt to do so could result in
infinite recursion. Also, don't re-check for a value that we already
established is missing in !bad-output's caller.

Hash randomization meant that the test could either pass or trigger the
infinite recursion; it seems it was written prior to hash randomization
and so it missed the logical bug described above.