Wire up filter groups to the formula language
This provides operators that are applied pairwise on the filter group
collection. One of them is used in the test model so far; we should add
some further ones too. Also exposed is a `scalar` builtin for turning a
filter group into its total value, and a `scale` builtin for multiplying
all filter group values by some scaling factor. Hopefully this is enough
to get a bit further with updating the model.