Provide a way to get all possible filter outputs
Keep track of the filter sets that contribute to a filter group in the
outputs. With the possible options in the enum attached to the filter
sets, this means we can produce all of the filter keys that could ever
occur. Use this to offer a means to get all of the filter group's values
including zero entries for those that have no instances using those enum
options. The options are sorted according to the enum within a given
module, however the order of modules is undetermined.