Refactor to prepare for multiple distributables
Switch the distribution hash to have the value be a list, not just a
single value. This means updating the code that produces it, and some
tweaks in the input distribution. It doesn't yet actually handle the
case of mutiple distributable inputs; this is just enough to get us able
to pass the (updated) existing tests for the case of a single input to
distribute, but with the tweaked data structures.