Implement input validation against the model
This catches a wide range of problems, including:

* Inputs for modules that don't exist
* Single-instance inputs for multi-instance modules and vice versa
* Missing inputs that don't have a default_calc
* Inputs whose value does not match the type
* Inputs for enum types where the value does not fall within the enum
* Inputs with range constraints (`between`, `ge`) that are not met by
  the value

For multi-instance modules, the instance at fault is also reported in
the error. A list of error objects is produced. This is not yet in any
way integrated with the overall workflow. Since we weren't validating
these things before, it's possible that starting to do so will turn up
genuine mistakes in input data. It's also possible more realistic data
will turn up problems in this validator, although hopefully the decent
level of test coverage here will mean we don't see too much of that.