Refactor application of default values
The primary motivation here is that we would like to have the defaults
applied *before* we peform branching and flattening, which menat that
the previous approach of looking at the defaults only when we needed to
access the input would not work. Given we'll also be adding support for
formulas to have a calculated default, this also makes sense, since we
can ensure we only evaluate that formula once.